server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
pymupdf==1.23.8
aiohttp==3.9.1
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2

